Boilerplate (#13)
* chore(markdown): consistency

* libCLImate: modernise C/C++ boilerplate for 0.4.1-alpha42

Bring **libCLImate** up to Synesis C/C++ boilerplate standard: CI,
helper scripts, CMake packaging, automated tests, and documentation.
No public API or runtime behaviour changes.

CI and GitHub Actions:
* Added modular **ci.yml** with six-cell OS/compiler matrix (Linux/macOS
  GCC/Clang, Windows MSVC, Windows MinGW) plus install-smoke job;
* Added reusable **ci-cell.yml** for configure/build/CTest and helper-
  script coverage (examples, unit, component, scratch);
* Added **install-sis-deps** composite action (verbatim **xTests**
  pattern); exports **SIS_DEPS** for callers;
* Build cells install full dependency chain: **STLSoft**, **b64**,
  **shwild**, **xTests**, **CLASP**, **Pantheios**,
  **Pantheios.Extras.DiagUtil**, **Pantheios.Extras.Main**;
* Install-smoke builds/installs with minimal deps, verifies install tree,
  and exercises **find_package(libCLImate)** consumer compile;
* **ctest --no-tests=error** gates empty-test regressions;
* Canonical push branches; unfiltered **pull_request**; **contents:
  read** permissions; concurrency with cancel-in-progress.

Helper scripts and project identity:
* Added **.sis/project_name.txt**; all scripts load **ProjectName** for
  status and **--help**;
* Added **ctest_cmake.sh**, **run_all_automated_tests.sh**, and
  **run_all_component_tests.sh**;
* Hardened all bash helpers: quoted paths, **find -print0 | sort -z**,
  **SIS_CMAKE_BUILD_DIR** / **SIS_CMAKE_MAKE_COMMAND** honouring,
  explicit empty-category handling, Synesis EOF markers;
* **prepare_cmake.sh**: collapsed MinGW/default **cmake** invocations
  (generator-only divergence); added **--build-shared-libs**; **MSYSTEM**
  no longer implies MinGW generator — only **--mingw** selects it;
* **run_all_unit_tests.sh** is unit-only; component runner succeeds when
  none found.

Automated tests:
* Added **test.unit.version** (version macro assertions) and
  **test.unit.api** (header smoke test) via xTests;
* Registered CTest targets in **define_automated_test_program**;
* Renamed scratch **libver** → **test.scratch.libver** for discovery.

CMake packaging:
* Renamed internal target **libclimate.core** with **libCLImate::core**
  alias; avoids parent-project **core** collision;
* **PUBLIC** interface links: **CLASP::core**, **Pantheios.Extras.Main**,
  conditional **STLSoft::STLSoft** (package route only);
* **libCLImate-config.cmake.in**: **find_dependency()** for true interface
  deps; conditional **STLSoft** guard via **STLSoft_IS_PACKAGE_**;
* Fixed **Pantheios.Extras.DiagUtil** **find_package()** version-variable
  typo;
* Fixed **INSTALL_DESTINATION** transposition for package config;
* **cmake/BuildType.cmake**: corrected **CMAKE_BUILD_TYPE** cache set;
* **cmake/LanguageFullVersion.cmake**: added **X_CMAKE_C_FULLSTANDARD**
  discrimination.

Documentation:
* **README.md**: tagline, CI badge row, filled Components (C/C++ API),
  corrected dependencies and related-project links;
* **CHANGES.md**: retitled; version-first headings; **0.4.1-alpha42**
  entry; merged **HISTORY.md** facts (git-tag-sourced **0.3.x** bullets);
* Removed **HISTORY.md**;
* **INSTALL.md**: fixed requisite list, corrupt steps, and consumer
  **find_package** example;
* **EXAMPLES.md**: synced table with real example directories;
* **NEWS.md**: added **0.4.1** row; corrected **0.3.6** date (24th January
  2017, per tag);
* **FAQ.md**: scaffold with EOF marker;
* **TODO.md**: marked packaging-improvement items complete.

Editor / tooling (incidental):
* **.gitattributes**: lexicographic revisibility; **CMakeLists.txt**
  linguist-detectable=false;
* **.vimrc** / **.vscode/settings.json**: tab/indent alignment and C++
  IntelliSense associations.

* fix: restore CI green paths and require CLASP 0.15.1.

Install Pantheios before Pantheios.Extras.Main in install-smoke; gate the
Windows wmain example behind LIBCLIMATE_USE_wmain; use a MinGW-safe size_t
format in examples/C/minimal; raise the CLASP floor to 0.15.1 (CMake,
README, and clasp.clasp.h guard).

* fix: install CMake package config under lowercase libclimate.

Use PROJECT_NAME_LOWER for EXPORT_NAME so find_package(libCLImate) finds
libclimate-config.cmake on case-sensitive filesystems, and align the
install-smoke verify path.

* fix: rename CMake package config template to libclimate-config.cmake.in.

Complete the EXPORT_NAME lowercase change with a case-only rename so
configure_package_config_file finds the template on case-sensitive
filesystems.

+++ examples/C/minimal/main.c
@@ -18,7 +18,7 @@ libCLImate_program_main(
     clasp_arguments_t const* args
 )
 {
-    pantheios_logprintf(PANTHEIOS_SEV_NOTICE, "args->numArguments=%zu", args->numArguments);
+    pantheios_logprintf(PANTHEIOS_SEV_NOTICE, "args->numArguments=%lu", (unsigned long)args->numArguments);
 
     return 0;
 }
